+++ CHANGELOG.md
@@ -3692,3 +3692,4 @@
 - [2026-08-18 11:39] perf(network): optimize memory consumption in network list viewport
 - [2026-08-18 11:48] test(i18n): implement widget integration tests for i18n
 - [2026-08-18 12:13] test(ui): add unit tests for ui business logic
+- [2026-08-18 12:50] feat(network): implement shimmer loading skeleton in network